Extract logical functions
